+++ src/main/java/tn/esprit/back/Services/Marketplace/IItemService.java
@@ -6,12 +6,12 @@ import java.util.List;
 import java.util.Optional;
 
 public interface IItemService {
-    Item ajouterItem(Item item);
+    // Retirer l'email comme paramètre ici et le gérer dans la méthode d'implémentation
+    Item ajouterItem(Item item);  // Ne prend plus userEmail en paramètre
     Item updateItem(Item item);
     List<Item> getAllItems();
     Optional<Item> getItemById(Long id);
     void supprimerItem(Long id);
 
-    List<Item> getItemsBySeller(Long sellerId);
-
+    List<Item> getItemsBySeller(int sellerId);
 }
